+++ R/pattern.R
@@ -8,26 +8,26 @@
 #' str(all_patterns)
 all_patterns = list(
   `rnw` = list(chunk.begin = '^\\s*<<(.*)>>=', chunk.end = '^\\s*@\\s*(%+.*|)$',
-               inline.code = '\\\\Sexpr\\{([^}]*)\\}',
+               inline.code = '\\\\Sexpr\\{([^}]+)\\}',
                inline.comment = '^\\s*%.*',
                ref.chunk = '^\\s*<<(.*)>>\\s*$',
                header.begin = '\\s*\\\\documentclass[^}]+\\}',
                document.begin = '\\s*\\\\begin\\{document\\}'),
 
-  `brew` = list(inline.code = '<%[=]{0,1}\\s+([^%]*)\\s+[-]*%>'),
+  `brew` = list(inline.code = '<%[=]{0,1}\\s+([^%]+)\\s+[-]*%>'),
 
   `tex` = list(chunk.begin = '^\\s*%+\\s*begin.rcode\\s*(.*)',
                chunk.end = '^\\s*%+\\s*end.rcode', chunk.code = '^%+',
                ref.chunk = '^%+\\s*<<(.*)>>\\s*$',
                inline.comment = '^\\s*%.*',
-               inline.code = '\\\\rinline\\{([^}]*)\\}',
+               inline.code = '\\\\rinline\\{([^}]+)\\}',
                header.begin = '\\s*\\\\documentclass[^}]+\\}',
                document.begin = '\\s*\\\\begin\\{document\\}'),
 
   `html` = list(chunk.begin = '^\\s*<!--\\s*begin.rcode\\s*(.*)',
                 chunk.end = '^\\s*end.rcode\\s*-->',
                 ref.chunk = '^\\s*<<(.*)>>\\s*$',
-                inline.code = '<!--\\s*rinline(.*?)-->',
+                inline.code = '<!--\\s*rinline(.+?)-->',
                 header.begin = '\\s*<head>'),
 
   `md` = list(chunk.begin = '^\\s*`{3,}\\s*\\{r(.*)\\}\\s*$',
@@ -39,7 +39,7 @@ all_patterns = list(
                chunk.end = "^\\s*\\.{2}\\s+\\.{2,}\\s*$",
                chunk.code = "^\\.{2}",
                ref.chunk = "^\\.*\\s*<<(.*)>>\\s*$",
-               inline.code = ":r:`([^`]*)`")
+               inline.code = ":r:`([^`]+)`")
 )
 
 .sep.label = '^#+\\s*@knitr(.*)$'  # pattern for code chunks in an R script