+++ inst/tests/test-parser.R
@@ -21,6 +21,9 @@ test_that('parsing inline texts', {
                        'another expression \\Sexpr{rnorm(10)}'), all_patterns$rnw$inline.code, NULL)
   expect_identical(res$code, c('x', 'NA', '1+2', 'rnorm(10)'))
   expect_identical(nchar(res$input), 81L)
+  # empty inline code is not recognized
+  expect_identical(parse_inline('\\Sexpr{}', all_patterns$rnw$inline.code)$code,
+                   character(0))
   # can use > in HTML inline code
   expect_identical(parse_inline('<!--rinline "<a>" -->', all_patterns$html$inline.code)$code,
                    ' "<a>" ')